Move filesize update job to a8c-files
We want the job to run on all environments regardless of wether they have Stream Wrappers activated or not
